Don't throw an error if no records are returned
If the code fails the twitter handshake dance it’ll _think_ there is a
use ID (in the session) but fail to pull it out of the database. User
will be `undefined`, however passport can’t handle that and will throw
an error. Setting `user` to `null` allows passport to gracefully carry
on.