add microformats markup to event pages
I tried to change as little about the markup as possible. No CSS changes were necessary, and this PR should not have any visible changes. There were a couple cases where I had to add an additional element to encompass some other elements on the page.

One pattern I added in a couple places has to do with profile photos for users. I added alt text with the user's name for the image, and wrapped the image in a link to the user's profile page. I hope that's okay.